fix(homeassistant): show "inactive" for bomb when value is null instead of "unknown"
When a value is null, home assistant will interpret the state as "Unknown", which looks odd. 
The bomb is always null when it's not planted, defused or exploded.
To fix this, we are going to create a value template that shows Inactive, instead of Unknown, which is a slightly better description.